Rename file proxy -> oauth_proxy
